mcp-client/requirements.txt
fastmcp
google-genai
